Add for user overrides
Allows the user/developer to set the defaults for the page via $.fn.fileReader.defaults. Very useful for setting things such as the filereader swf path which should be the same for all instances on the page.
Also, detects if native version exists in browser and removes polyfill code if not needed.